+++ example/lib/main.dart
@@ -1,0 +1,11 @@
+import 'package:example/src/configuration/default_application_configuration.dart';
+import 'package:flutter/material.dart';
+import 'package:fappconfiguration/fappconfiguration.dart';
+
+void main() {
+  runApp(
+    DefaultMaterialApplication(
+      DefaultApplicationConfiguration(),
+    ),
+  );
+}