+++ lib/fappconfiguration.dart
@@ -1,6 +1,7 @@
 library fappconfiguration;
 
-export "src/application_configuration.dart";
-export "src/default_cupertino_application.dart";
-export "src/default_material_application.dart";
-export "src/dynamic_widget_provider.dart";
+export 'src/application_configuration.dart';
+export 'src/default_cupertino_application.dart';
+export 'src/default_material_application.dart';
+export 'src/dynamic_widget_provider.dart';
+export 'src/fappconfiguration_application.dart';