+++ lib/src/default_cupertino_application.dart
@@ -1,43 +1,71 @@
 import "package:fappconfiguration/src/application_configuration.dart"
     show ApplicationConfiguration;
+import 'package:fappconfiguration/src/fappconfiguration_application.dart';
 import "package:flutter/cupertino.dart";
 import "package:flutter_localizations/flutter_localizations.dart"
     show GlobalCupertinoLocalizations, GlobalWidgetsLocalizations;
-import "package:provider/provider.dart" show Consumer, Provider;
+import "package:provider/provider.dart" show Provider, MultiProvider;
 
-/// Default iOS Cupertino Application that configure a [CupertinoApp] using the
-/// [ApplicationConfiguration].
-class DefaultCupertinoApplication extends StatelessWidget {
+/// Default iOS Cupertino Application that configure a [CupertinoApp]
+/// using the properties of an [ApplicationConfiguration].
+class DefaultCupertinoApplication extends FAppConfigurationApplication {
   /// Construct a Cupertino application using [ApplicationConfiguration].
-  const DefaultCupertinoApplication(this._configuration, {Key key})
-      : assert(_configuration != null, "configuration can't be null"),
-        super(key: key);
-
-  final ApplicationConfiguration _configuration;
+  const DefaultCupertinoApplication(
+    final ApplicationConfiguration initialConfiguration, {
+    Key key,
+  })  : assert(
+          initialConfiguration != null,
+          "initialConfiguration can't be null",
+        ),
+        super(initialConfiguration, key: key);
 
   @override
+  _DefaultCupertinoApplicationState createState() =>
+      _DefaultCupertinoApplicationState();
+}
+
+class _DefaultCupertinoApplicationState
+    extends FAppConfigurationApplicationState {
+  @override
   Widget build(BuildContext context) {
-    return Provider.value(
-      value: _configuration,
-      child: Consumer<ApplicationConfiguration>(builder: (context, value, _) {
-        return CupertinoApp(
-          title: value.name,
-          theme: CupertinoThemeData(
-            primaryColor: value.theme.primaryColor,
-            primaryContrastingColor: value.theme.accentColor,
-            brightness: value.theme.brightness,
-            textTheme: CupertinoTextThemeData(
-              navTitleTextStyle: value.theme.textTheme.title,
-            ),
-          ),
-          supportedLocales: value.supportedLanguages,
-          localizationsDelegates: value.localizationsDelegates
-            ..add(GlobalCupertinoLocalizations.delegate)
-            ..add(GlobalWidgetsLocalizations.delegate),
-          onGenerateRoute: value.routeFactory,
-          onUnknownRoute: value.unknownRouteFactory,
-        );
-      }),
+    return MultiProvider(
+      providers: [
+        Provider.value(value: configuration),
+        ...?configuration.dependencies,
+      ],
+      child: createIosAppWidget(context),
+    );
+  }
+
+  CupertinoApp createIosAppWidget(BuildContext context) {
+    final platformBrightness = MediaQuery.platformBrightnessOf(context);
+
+    final currentTheme = platformBrightness == Brightness.light
+        ? configuration.theme
+        : configuration.darkTheme;
+
+    return CupertinoApp(
+      navigatorKey: configurationKey,
+      title: configuration.name,
+      theme: CupertinoThemeData(
+        primaryColor: currentTheme.primaryColor,
+        primaryContrastingColor: currentTheme.accentColor,
+        brightness: currentTheme.brightness,
+        textTheme: CupertinoTextThemeData(
+          brightness: currentTheme.brightness,
+          navTitleTextStyle: currentTheme.primaryTextTheme.title,
+        ),
+        scaffoldBackgroundColor: currentTheme.scaffoldBackgroundColor,
+        barBackgroundColor: currentTheme.bottomAppBarColor,
+      ),
+      supportedLocales: configuration.supportedLanguages,
+      localizationsDelegates: [
+        ...?configuration.localizationsDelegates,
+        GlobalCupertinoLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+      ],
+      onGenerateRoute: configuration.routeFactory,
+      onUnknownRoute: configuration.unknownRouteFactory,
     );
   }
 }